Some rearranging.
