+++ setup.py
@@ -5,6 +5,7 @@ setup(name="quartic-python",
       version=os.environ.get("CIRCLE_BUILD_NUM", "unknown"),
       description="Quartic platform bindings",
       author="Quartic Technologies",
+      version="0.1",
       author_email="contact@quartic.io",
       license="BSD",
       packages=find_packages("src"),
